mid-refactoring for NxM node vs. client

+++ src/libs/libcore/widgets/ArduinoPinMap.cpp
@@ -144,7 +144,7 @@ QStringList ArduinoPinMap::pinStringList()
 
 ArduinoPin ArduinoPinMap::pin(quint32 idx)
 {
-	if(idx>=mPins.size()) {
+	if(idx>=(quint32)mPins.size()) {
 		return ArduinoPin();
 	}
 	return mPins[idx];
